functions/requirements.txt
google-cloud-secret-manager==2.0.0
google-cloud-storage==2.1.0
gkeepapi==0.13.7
